Added RestTemplate

+++ src/main/java/com/devlopment/bookstore_cartservices/service/CartService.java
@@ -1,6 +1,7 @@
 package com.devlopment.bookstore_cartservices.service;
 
 import com.devlopment.bookstore_cartservices.dto.CartRequestDTO;
+import com.devlopment.bookstore_cartservices.dto.UserResponseDTO;
 import com.devlopment.bookstore_cartservices.entity.Cart;
 import com.devlopment.bookstore_cartservices.entity.UserData;
 import com.devlopment.bookstore_cartservices.exception.InvalidTokenException;
@@ -8,67 +9,78 @@ import com.devlopment.bookstore_cartservices.repository.CartRepository;
 import com.devlopment.bookstore_cartservices.util.TokenUtility;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
 import java.util.Optional;
 @Service
 public class CartService implements ICartService {
     @Autowired
     private CartRepository cartRepository;
+    @Autowired
+    TokenUtility tokenUtility;
+    @Autowired
+    RestTemplate restTemplate;
     @Override
-    public Cart createCart(CartRequestDTO cartRequestDTO) {
-        /*int userId= tokenUtility.decodeToken(token);
-        Optional<UserData> optionalUserData = userRepository.findById(userId);
-        if(optionalUserData.isPresent())
-        {*/
+    public Cart createCart(CartRequestDTO cartRequestDTO,String token) {
+        int userId= tokenUtility.decodeToken(token);
+        UserResponseDTO userResponseDTO= restTemplate.getForObject("http://localhost:8086/userservice/" + userId,UserResponseDTO.class);
+        UserData userData = userResponseDTO.getUserData();
+        if(userData!= null)
+        {
             Cart cart =new Cart();
-            //UserData loggedInUserData = optionalUserData.get();
-           // cart.setUserData(loggedInUserData);
+            UserData loggedInUserData = userResponseDTO.getUserData();
+            cart.setUserData(loggedInUserData);
+            cart.setQuantity(cartRequestDTO.getQuantity());
+            cart.setBookIds(cartRequestDTO.getBookIds());
             return cartRepository.save(cart);
-        /*}
+        }
         else {
             throw new InvalidTokenException(token);
-        }*/
+        }
     }
 
     @Override
-    public Cart getCart(int cartId) {
-        /*int userId= tokenUtility.decodeToken(token);
-        Optional<UserData> optionalUserData = userRepository.findById(userId);
-        if(optionalUserData.isPresent())
-        {*/
+    public Cart getCart(int cartId,String token) {
+        int userId= tokenUtility.decodeToken(token);
+        UserResponseDTO userResponseDTO= restTemplate.getForObject("http://localhost:8086/userservice/" + userId,UserResponseDTO.class);
+        UserData userData = userResponseDTO.getUserData();
+        if(userData!= null)
+        {
             return cartRepository.findById(cartId).get();
-       /* }
+        }
         else {
             throw new InvalidTokenException(token);
-        }*/
+        }
     }
 
     @Override
-    public Cart updateCart(int cartId, CartRequestDTO cartRequestDTO) {
-        /*int userId= tokenUtility.decodeToken(token);
-        Optional<UserData> optionalUserData = userRepository.findById(userId);
-        if(optionalUserData.isPresent())
-        {*/
+    public Cart updateCart(int cartId, CartRequestDTO cartRequestDTO,String token) {
+        int userId= tokenUtility.decodeToken(token);
+        UserResponseDTO userResponseDTO= restTemplate.getForObject("http://localhost:8086/userservice/" + userId,UserResponseDTO.class);
+        UserData userData = userResponseDTO.getUserData();
+        if(userData!= null)
+        {
             Cart cart=cartRepository.getReferenceById(cartId);
             cart.setQuantity(cartRequestDTO.getQuantity());
-            //cart.setBookIds(cartRequestDTO.getBookIds());
+            cart.setBookIds(cartRequestDTO.getBookIds());
             return cartRepository.save(cart);
-      /*  }
+       }
         else {
             throw new InvalidTokenException(token);
-        }*/
+        }
     }
 
     @Override
-    public void deleteCart(int cartId) {
-        /*int userId= tokenUtility.decodeToken(token);
-        Optional<UserData> optionalUserData = userRepository.findById(userId);
-        if(optionalUserData.isPresent())
-        {*/
+    public void deleteCart(int cartId,String token) {
+        int userId= tokenUtility.decodeToken(token);
+        UserResponseDTO userResponseDTO= restTemplate.getForObject("http://localhost:8086/userservice/" + userId,UserResponseDTO.class);
+        UserData userData = userResponseDTO.getUserData();
+        if(userData!= null)
+        {
             cartRepository.deleteById(cartId);
-        /*}
+        }
         else {
             throw new InvalidTokenException(token);
-        }*/
+        }
     }
 }
